Implemented the Hall A Moller Quads (Q1,Q2,Q3).
The positions and dimensions of the Quads are described in the GDML
files. The magnetic field gradient is determined from the equations in
http://hallaweb.jlab.org/equipment/moller/magnets/Moller_quads1.pdf

In order to fit the magnets in, I had to comment out both Dummy
detectors. Also, made the world be under Vacuum in order to remove all
the multiple scattering with the air. The whole beamline would be under
vacuum anyway.